Title: 합과 차의 나눗셈, Time: 6ms, Memory: 0MB, Status: Wrong Answer - CodeTree

+++ 241001/합과 차의 나눗셈/divide-of-sum-and-sub.cpp
@@ -4,10 +4,10 @@ using namespace std;
 int main() {
     int a, b;
     cin >> a >> b;
-    int sum = a+b;
-    int sub = a-b;
+    double sum = a+b;
+    double sub = a-b;
 
     cout.precision(3);
-    cout << double(sum)/sub;
+    cout << sum/sub;
     return 0;
 }
